split math from std
